requirements.txt
setuptools>=41.0.0
wheel~=0.35
tensorflow==2.1.0
git+https://github.com/GOTO-OBS/catsHTM.git#egg=catsHTM # needed for fixed catsHTM
absl-py==0.8.0
asn1crypto==1.0.0
astor==0.8.0
astropy==3.2.1
backcall==0.1.0
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
dill==0.3.1.1
ephem==3.7.7.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
hickle==3.4.5
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
joblib==0.14.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mock==3.0.5
numexpr==2.7.0
numpy==1.17.2
opt-einsum==3.1.0
pandas~=1.0.5
paramiko==2.6.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pillow==7.0.0
prompt-toolkit==2.0.10
protobuf==3.10.0
psycopg2-binary==2.8.3
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pympc>=0.6.1
PyNaCl==1.3.0
pyparsing==2.4.2
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
scikit-learn>=0.22.0
scipy~=1.4.1
six==1.12.0
tables~=3.6.1
tensorboard~=2.1.0
tensorflow-estimator~=2.1.0
termcolor==1.1.0
traitlets==4.3.3
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
tqdm==4.43.0
keras-tuner~=1.0.0
